Add Data Access Objects for contracts, services, users, and available dates
- Created ContratacaoDAO for managing contracts, including creation, updates, and statistics.
- Implemented DataDisponivelDAO for handling available dates for services.
- Added ServicoDAO for service management, including CRUD operations and searching.
- Introduced UsuarioDAO for user authentication and management.
- Created a generic Conexao class for database connection handling.
- Enhanced error logging for better debugging and tracking of issues.
