+++ admin/clientes/form.php
@@ -1,6 +1,6 @@
 <?php
-require_once '../../config/auth.php';
-require_once '../../dao/ClienteDAO.php';
+require_once '../../shared/config/auth.php';
+require_once '../../shared/dao/ClienteDAO.php';
 
 Auth::requireAuth();
 
@@ -135,7 +135,7 @@ function isValidCPF($cpf)
     <title><?= $isEdit ? 'Editar' : 'Novo' ?> Cliente - Sistema de Gestão de Serviços</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css" rel="stylesheet">
-    <link rel="stylesheet" href="../assets/css/style.css">
+    <link rel="stylesheet" href="../../shared/assets/css/style.css">
 </head>
 
 <body>
@@ -185,7 +185,7 @@ function isValidCPF($cpf)
                             <i class="bi bi-person-circle"></i> <?= htmlspecialchars($user['nome']) ?>
                         </a>
                         <ul class="dropdown-menu">
-                            <li><a class="dropdown-item" href="logout.php">
+                            <li><a class="dropdown-item" href="../logout.php">
                                     <i class="bi bi-box-arrow-right"></i> Sair
                                 </a></li>
                         </ul>
@@ -303,7 +303,7 @@ function isValidCPF($cpf)
 
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="../assets/js/util.js"></script>
+    <script src="../../shared/assets/js/util.js"></script>
     <script>
         $(document).ready(function() {
             // CPF mask and validation
