+++ admin/contratacao/carrinho.php
@@ -1,5 +1,5 @@
 <?php
-require_once '../../config/auth.php';
+require_once '../../shared/config/auth.php';
 
 Auth::requireAuth();
 
@@ -20,7 +20,7 @@ $cart = $_SESSION['cart'];
     <title>Carrinho de Compras - Sistema de Gestão de Serviços</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css" rel="stylesheet">
-    <link rel="stylesheet" href="../assets/css/style.css">
+    <link rel="stylesheet" href="../../shared/assets/css/style.css">
 </head>
 
 <body>
@@ -70,7 +70,7 @@ $cart = $_SESSION['cart'];
                             <i class="bi bi-person-circle"></i> <?= htmlspecialchars(Auth::getUser()['nome']) ?>
                         </a>
                         <ul class="dropdown-menu">
-                            <li><a class="dropdown-item" href="logout.php">
+                            <li><a class="dropdown-item" href="../logout.php">
                                     <i class="bi bi-box-arrow-right"></i> Sair
                                 </a></li>
                         </ul>
@@ -168,7 +168,7 @@ $cart = $_SESSION['cart'];
 
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="../assets/js/util.js"></script>
+    <script src="../../shared/assets/js/util.js"></script>
     <script src="carrinho.js"></script>
     <script>
         $(document).ready(function() {
