+++ admin/contratacao/listar.php
@@ -1,8 +1,8 @@
 <?php
-require_once '../../config/auth.php';
-require_once '../../dao/ContratacaoDAO.php';
-require_once '../../dao/ClienteDAO.php';
-require_once '../../dao/ServicoDAO.php';
+require_once '../../shared/config/auth.php';
+require_once '../../shared/dao/ContratacaoDAO.php';
+require_once '../../shared/dao/ClienteDAO.php';
+require_once '../../shared/dao/ServicoDAO.php';
 
 Auth::requireAuth();
 
@@ -44,7 +44,7 @@ $totalReceita = array_sum(array_column($contratos, 'valor_total'));
     <title>Gerenciar Contratos - Sistema de Gestão</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css" rel="stylesheet">
-    <link rel="stylesheet" href="../assets/css/style.css">
+    <link rel="stylesheet" href="../../shared/assets/css/style.css">
 </head>
 
 <body>
@@ -381,7 +381,7 @@ $totalReceita = array_sum(array_column($contratos, 'valor_total'));
 
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="../assets/js/util.js"></script>
+    <script src="../../shared/assets/js/util.js"></script>
 
     <script>
         function viewContract(contractId) {
