+++ admin/contratacao/resumo.php
@@ -1,7 +1,7 @@
 <?php
-require_once '../../config/auth.php';
-require_once '../../dao/ServicoDAO.php';
-require_once '../../dao/ClienteDAO.php';
+require_once '../../shared/config/auth.php';
+require_once '../../shared/dao/ServicoDAO.php';
+require_once '../../shared/dao/ClienteDAO.php';
 
 Auth::requireAuth();
 
@@ -58,7 +58,7 @@ $csrfToken = Auth::generateCSRFToken();
     <title>Resumo da Contratação - Sistema de Gestão de Serviços</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css" rel="stylesheet">
-    <link rel="stylesheet" href="../assets/css/style.css">
+    <link rel="stylesheet" href="../../shared/assets/css/style.css">
 </head>
 
 <body>
@@ -114,7 +114,7 @@ $csrfToken = Auth::generateCSRFToken();
                             <i class="bi bi-person-circle"></i> <?= htmlspecialchars($user['nome']) ?>
                         </a>
                         <ul class="dropdown-menu">
-                            <li><a class="dropdown-item" href="logout.php">
+                            <li><a class="dropdown-item" href="../logout.php">
                                     <i class="bi bi-box-arrow-right"></i> Sair
                                 </a></li>
                         </ul>
