+++ admin/contratacao/update_status.php
@@ -1,6 +1,6 @@
 <?php
-require_once '../../config/auth.php';
-require_once '../../dao/ContratacaoDAO.php';
+require_once '../../shared/config/auth.php';
+require_once '../../shared/dao/ContratacaoDAO.php';
 
 Auth::requireAuth();
