+++ admin/dashboard.php
@@ -1,8 +1,8 @@
 <?php
-require_once '../config/auth.php';
-require_once '../dao/ClienteDAO.php';
-require_once '../dao/ServicoDAO.php';
-require_once '../dao/ContratacaoDAO.php';
+require_once '../shared/config/auth.php';
+require_once '../shared/dao/ClienteDAO.php';
+require_once '../shared/dao/ServicoDAO.php';
+require_once '../shared/dao/ContratacaoDAO.php';
 
 Auth::requireAuth();
 
@@ -27,7 +27,7 @@ $user = Auth::getUser();
     <title>Dashboard - Sistema de Gestão de Serviços</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css" rel="stylesheet">
-    <link rel="stylesheet" href="../assets/css/style.css">
+    <link rel="stylesheet" href="../shared/assets/css/style.css">
 </head>
 
 <body>
@@ -244,7 +244,7 @@ $user = Auth::getUser();
 
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="../assets/js/util.js"></script>
+    <script src="../shared/assets/js/util.js"></script>
 </body>
 
 </html>
