+++ public/checkout.php
@@ -13,9 +13,9 @@ if (!isset($_SESSION['client_id'])) {
     exit;
 }
 
-require_once 'dao/ServicoDAO.php';
-require_once 'dao/ContratacaoDAO.php';
-require_once 'dao/DataDisponivelDAO.php';
+require_once '../shared/dao/ServicoDAO.php';
+require_once '../shared/dao/ContratacaoDAO.php';
+require_once '../shared/dao/DataDisponivelDAO.php';
 
 // Verificar se existe carrinho
 if (!isset($_SESSION['cart']) || empty($_SESSION['cart'])) {
@@ -84,7 +84,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     <title>Finalizar Pedido - Sistema de Serviços</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.0/font/bootstrap-icons.css" rel="stylesheet">
-    <link href="assets/css/style.css" rel="stylesheet">
+    <link href="../shared/assets/css/style.css" rel="stylesheet">
 </head>
 
 <body>
@@ -244,7 +244,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     <script src="https://code.jquery.com/jquery-3.7.1.min.js"></script>
-    <script src="admin/contratacao/carrinho.js"></script>
+    <script src="../admin/contratacao/carrinho.js"></script>
     <script>
         $(document).ready(function() {
             // Se checkout foi bem-sucedido, limpar carrinho do localStorage
