+++ public/login.php
@@ -7,7 +7,7 @@
 
 session_start();
 
-require_once 'dao/ClienteDAO.php';
+require_once '../shared/dao/ClienteDAO.php';
 
 $clienteDAO = new ClienteDAO();
 $error = '';
@@ -53,7 +53,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     <title>Login Cliente - Sistema de Serviços</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.0/font/bootstrap-icons.css" rel="stylesheet">
-    <link href="../assets/css/style.css" rel="stylesheet">
+    <link href="../shared/assets/css/style.css" rel="stylesheet">
 </head>
 
 <body class="bg-light">
@@ -71,7 +71,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                 <a class="nav-link" href="cadastro.php">
                     <i class="bi bi-person-plus"></i> Cadastrar
                 </a>
-                <a class="nav-link" href="admin/">
+                <a class="nav-link" href="../admin/">
                     <i class="bi bi-gear"></i> Admin
                 </a>
             </div>
