+++ public/logout_cliente.php
